requirements.txt
openai~=2.14.0
python-dotenv~=1.2.1
rich~=14.2.0
langdetect